+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 - Introduce experimental uncurried by default mode. Can be turned on mid-file by adding standalone annotation `@@uncurried`. For experimentation only. https://github.com/rescript-lang/rescript-compiler/pull/5796
 
+- Adding `@@toUncurried` to the file and reformat will convert to uncurried syntax
+
 #### :boom: Breaking Change
 
 - Remove support for the legacy Reason syntax. Existing Reason code can be converted to ReScript syntax using ReScript 9 as follows: